refactor: move scheduler to system_services + remove TaskQueueService
Phase 1: Move services/scheduler/ → system_services/scheduler/ to fix
KERNEL-ARCHITECTURE.md §3 tier violation (SchedulerProtocol is "Agent
Infra" tier, same as AgentRegistry). Rewrites ~40 imports across source
and test files.

Phase 2: Remove TaskQueueService entirely — zero runtime consumers,
overlaps with SchedulerService, adds unnecessary Rust build dependency.
Deletes service, protocol, factory wiring, lifespan init, config fields,
NexusFS RPC dispatch entries, RPC codegen ref, and all tests.

45 files changed, -752 lines deleted.
