+++ src/nexus/services/__init__.py
@@ -11,8 +11,6 @@ Services:
 - MCPService: Model Context Protocol server management
 - LLMService: LLM-powered document reading with citations
 - OAuthCredentialService: OAuth credential management (brick)
-- SchedulerService: Fair-share priority scheduler (Astraea) — System Service, not a Brick
-
 Subsystem ABC (Issue #1287):
 - Subsystem: ABC for service lifecycle wrappers (health_check, cleanup)
 - ContextIdentity: Frozen identity extracted from OperationContext
@@ -24,7 +22,6 @@ Phase 2: Core Refactoring (Issue #988)
 import importlib as _il
 
 from nexus.contracts.types import ContextIdentity, extract_context_identity
-from nexus.services.scheduler import SchedulerService
 from nexus.services.search.search_service import SearchService
 from nexus.services.subsystem import Subsystem
 from nexus.services.versioning.version_service import VersionService
@@ -45,7 +42,6 @@ __all__ = [
     "MCPService",
     "LLMService",
     "OAuthCredentialService",
-    "SchedulerService",
     "Subsystem",
     "ContextIdentity",
     "extract_context_identity",
