+++ src/nexus/system_services/agents/async_agent_registry.py
@@ -23,8 +23,8 @@ from nexus.contracts.protocols.agent_registry import AgentInfo
 
 if TYPE_CHECKING:
     from nexus.contracts.agent_types import AgentRecord, AgentSpec, AgentStatus
-    from nexus.services.scheduler.events import AgentStateEmitter
     from nexus.system_services.agents.agent_registry import AgentRegistry
+    from nexus.system_services.scheduler.events import AgentStateEmitter
 
 
 def _to_agent_info(record: "AgentRecord") -> AgentInfo:
@@ -122,7 +122,7 @@ class AsyncAgentRegistry:
 
         # Emit state change event if emitter is configured
         if self._state_emitter is not None and previous_state is not None:
-            from nexus.services.scheduler.events import AgentStateEvent
+            from nexus.system_services.scheduler.events import AgentStateEvent
 
             event = AgentStateEvent(
                 agent_id=agent_id,
