+++ src/nexus/system_services/lifecycle/__init__.py
@@ -7,12 +7,10 @@ from nexus.system_services.lifecycle.brick_lifecycle import BrickLifecycleManage
 from nexus.system_services.lifecycle.brick_reconciler import BrickReconciler
 from nexus.system_services.lifecycle.events_service import EventsService
 from nexus.system_services.lifecycle.expectations import Expectations
-from nexus.system_services.lifecycle.task_queue_service import TaskQueueService
 
 __all__ = [
     "BrickLifecycleManager",
     "BrickReconciler",
     "EventsService",
     "Expectations",
-    "TaskQueueService",
 ]
