+++ src/nexus/system_services/scheduler/dispatcher.py
@@ -14,13 +14,13 @@ import contextlib
 import logging
 from typing import TYPE_CHECKING, Any
 
-from nexus.services.scheduler.constants import (
+from nexus.system_services.scheduler.constants import (
     AGING_INTERVAL_SECONDS,
     STARVATION_PROMOTION_THRESHOLD_SECS,
 )
 
 if TYPE_CHECKING:
-    from nexus.services.scheduler.service import SchedulerService
+    from nexus.system_services.scheduler.service import SchedulerService
 
 logger = logging.getLogger(__name__)
