+++ tests/benchmarks/bench_factory_boot.py
@@ -357,7 +357,6 @@ class TestPerTierBreakdown:
             "workflow_engine",
             "api_key_creator",
             "snapshot_service",
-            "task_queue_service",
             "ipc_storage_driver",
             "ipc_vfs_driver",
             "ipc_provisioner",
