+++ tests/benchmarks/test_scheduler_benchmarks.py
@@ -8,9 +8,9 @@ import time
 
 import pytest
 
-from nexus.services.scheduler.constants import PriorityTier, RequestState
-from nexus.services.scheduler.policies.classifier import classify_request
-from nexus.services.scheduler.policies.hrrn import compute_hrrn_score, rank_by_hrrn
+from nexus.system_services.scheduler.constants import PriorityTier, RequestState
+from nexus.system_services.scheduler.policies.classifier import classify_request
+from nexus.system_services.scheduler.policies.hrrn import compute_hrrn_score, rank_by_hrrn
 
 
 @pytest.mark.benchmark
