+++ tests/e2e/self_contained/test_scheduler_integration.py
@@ -21,16 +21,16 @@ import pytest
 
 from nexus.bricks.pay.credits import CreditsService
 from nexus.contracts.protocols.scheduler import AgentRequest
-from nexus.services.scheduler.constants import (
+from nexus.system_services.scheduler.constants import (
     AGING_THRESHOLD_SECONDS,
     BOOST_COST_PER_TIER,
     TASK_STATUS_QUEUED,
     TASK_STATUS_RUNNING,
     PriorityTier,
 )
-from nexus.services.scheduler.models import ScheduledTask, TaskSubmission
-from nexus.services.scheduler.priority import compute_effective_tier
-from nexus.services.scheduler.service import SchedulerService
+from nexus.system_services.scheduler.models import ScheduledTask, TaskSubmission
+from nexus.system_services.scheduler.priority import compute_effective_tier
+from nexus.system_services.scheduler.service import SchedulerService
 
 # =============================================================================
 # Fixtures
