+++ tests/unit/scheduler/test_astraea_service.py
@@ -10,15 +10,15 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 
 from nexus.contracts.protocols.scheduler import AgentRequest
-from nexus.services.scheduler.constants import (
+from nexus.system_services.scheduler.constants import (
     TASK_STATUS_RUNNING,
     PriorityClass,
     PriorityTier,
 )
-from nexus.services.scheduler.events import AgentStateEmitter, AgentStateEvent
-from nexus.services.scheduler.models import ScheduledTask
-from nexus.services.scheduler.policies.fair_share import FairShareCounter
-from nexus.services.scheduler.service import SchedulerService
+from nexus.system_services.scheduler.events import AgentStateEmitter, AgentStateEvent
+from nexus.system_services.scheduler.models import ScheduledTask
+from nexus.system_services.scheduler.policies.fair_share import FairShareCounter
+from nexus.system_services.scheduler.service import SchedulerService
 
 # =============================================================================
 # Fixtures
