+++ tests/unit/scheduler/test_fair_share.py
@@ -5,7 +5,7 @@ Tests admission, counters, limits, sync, and snapshots.
 
 import pytest
 
-from nexus.services.scheduler.policies.fair_share import FairShareCounter, FairShareSnapshot
+from nexus.system_services.scheduler.policies.fair_share import FairShareCounter, FairShareSnapshot
 
 
 class TestFairShareSnapshot:
