+++ tests/unit/scheduler/test_hrrn.py
@@ -7,7 +7,7 @@ import pytest
 from hypothesis import given, settings
 from hypothesis import strategies as st
 
-from nexus.services.scheduler.policies.hrrn import compute_hrrn_score, rank_by_hrrn
+from nexus.system_services.scheduler.policies.hrrn import compute_hrrn_score, rank_by_hrrn
 
 
 class TestComputeHrrnScore:
